added load button and ability to load songs, also via pickle
A big issue arises that I had planned to fix later: pickle can't serialize the pygame.mixer in a sound object so I just commented it out. I would like to move the sound playing to the controller anyways, I think I can make it faster that way. Now its required though because as is a sound cannot play.